+++ src/extension.ts
@@ -39,24 +39,6 @@ export function activate(context: vscode.ExtensionContext) {
         })
     )
 
-    context.subscriptions.push(
-        vscode.commands.registerCommand('gears.startRuntime', () => {
-            const version = config("gears.runtime.version")
-            if (!version) {
-                vscode.window.showErrorMessage('The GEARS Runtime version is not configured')
-                return
-            }
-            const image = "xlrit/gears-runtime" + version
-            vscode.window.showInformationMessage(`Running ${image} is not supported yet`)
-            // zie https://github.com/microsoft/vscode-docker/blob/master/src/utils/executeAsTask.ts
-            // of iets als executeCommand vscode-docker.containers.start
-            // en dan verschillende opties:
-            // - gears runtime run --image ${image}
-            // - docker run -d --rm -p 1110:110 -p 2525:25 -p 8080:8080 -p 9990:9990 --name gears-runtime ${image}
-            // - docker-compose up
-        })
-    )
-
     console.log('Extension "vscode-gears" is now active')
 }