+++ src/gearsTaskProvider.ts
@@ -36,15 +36,13 @@ export class GearsTaskProvider implements vscode.TaskProvider {
             return task
         }
         return [
-            createTask('0. Start Runtime', this.startRuntimeExecution()),
-            createTask('1. Generate',      this.generateExecution()),
-            createTask('2. Diagrams',      this.diagramsExecution()),
-            createTask('3. Build',         this.buildExecution()),
-            createTask('4. Deploy',        this.deployExecution()),
-            createTask('5. Load data',     this.loadDataExecution()),
-            createTask('6. Run scenarios', this.runScenariosExecution()),
-            createTask('7. Undeploy',      this.undeployExecution()),
-            createTask('8. Stop Runtime',  this.stopRuntimeExecution()),
+            createTask('1. Generate',          this.generateExecution()),
+            createTask('2. Show Diagrams',     this.diagramsExecution()),
+            createTask('3. Build',             this.buildExecution()),
+            createTask('4. Run Application',   this.runExecution()),
+            createTask('5. Load data',         this.loadDataExecution()),
+            createTask('6. Run scenarios',     this.runScenariosExecution()),
+            //createTask('7. Stop Application',  this.stopExecution()),
         ];
     }
 
@@ -53,43 +51,6 @@ export class GearsTaskProvider implements vscode.TaskProvider {
         const newTask = new vscode.Task(def, task.scope, def.name, def.type, task.execution);
         newTask.group = task.group;
         return newTask;
-    }
-
-    startRuntimeExecution(): Execution {
-        const cwd = this.workspaceRoot
-        const cmd = this.startRuntimeCmd()
-        return new vscode.ShellExecution(cmd, { cwd })
-    }
-
-    startRuntimeCmd(): string {
-        const runtimeVersion = this.config('runtime.version')
-        switch (this.config('runtime.management.mode')) {
-            case 'gears-cli': 
-                return `gears runtime run --version ${runtimeVersion}`
-            case 'docker-compose':
-                return 'docker-compose up'
-            default: // docker
-                const docker = this.config('docker')
-                return `${docker} run -d --rm -p 1110:110 -p 2525:25 -p 8080:8080 -p 9990:9990 --name gears-runtime xlrit/gears-runtime:v${runtimeVersion}`
-        }
-    }
-
-    stopRuntimeExecution(): Execution {
-        const cwd = this.workspaceRoot
-        const cmd = this.stopRuntimeCmd()
-        return new vscode.ShellExecution(cmd, { cwd })
-    }
-
-    stopRuntimeCmd(): string {
-        switch (this.config("runtime.management.mode")) {
-            case 'gears-cli': 
-                return `gears runtime kill`
-            case 'docker-compose':
-                return 'docker-compose down'
-            default: // docker
-                const docker = this.config('docker')
-                return `${docker} kill gears-runtime`
-        }
     }
 
     generateExecution(): Execution {
@@ -105,7 +66,7 @@ export class GearsTaskProvider implements vscode.TaskProvider {
         if (projectName)    cmd += ` --name ${projectName}`
         if (projectVersion) cmd += ` --version ${projectVersion}`
         if (runtimeVersion) cmd += ` --runtime-version ${runtimeVersion}`
-        if (extraArgs) cmd += ` ${extraArgs}`
+        if (extraArgs)      cmd += ` ${extraArgs}`
         cmd += ` ${filePattern}`
         return new vscode.ShellExecution(cmd, { cwd })
     }
@@ -130,47 +91,18 @@ export class GearsTaskProvider implements vscode.TaskProvider {
         return new vscode.ShellExecution(cmd, { cwd })
     }
 
-    deployExecution(): Execution {
-        const projectName    = this.config('project.name')
-        const projectVersion = this.config('project.version')
-        const cwd  = `${this.workspaceRoot}/target/${projectName}`
-        const cmd  = this.getDeployCmd('target', `${projectName}-${projectVersion}.war`)
-        return new vscode.ShellExecution(cmd,  { cwd })
+    runExecution(): Execution {
+        const projectName = this.config('project.name')
+        const cwd = `${this.workspaceRoot}/target/${projectName}`
+        const cmd = 'mvn spring-boot:run'
+        return new vscode.ShellExecution(cmd, { cwd })
     }
 
-    getDeployCmd(dir: string, file: string): string {
-        const path = `${dir}/${file}`
-        switch (this.config("deploy.mode")) {
-            case 'gears-cli': 
-                return `gears runtime deploy ${path}`
-            case 'maven':
-                return 'mvn org.wildfly.plugins:wildfly-maven-plugin:2.0.2.Final:deploy'
-            default:
-                const docker = this.config('docker')
-                return `${docker} cp ${path} gears-runtime:/camunda/standalone/deployments && ` +
-                       `${docker} exec -it gears-runtime wait-for-deployment ${file}`
-        }
-    }
-
-    undeployExecution(): Execution {
-        const projectName    = this.config('project.name')
-        const projectVersion = this.config('project.version')
-        const cwd  = `${this.workspaceRoot}/target/${projectName}`
-        const cmd  = this.getUndeployCmd('target', `${projectName}-${projectVersion}.war`)
-        return new vscode.ShellExecution(cmd,  { cwd })
-    }
-
-    getUndeployCmd(dir: string, file: string): string {
-        const path = `${dir}/${file}`
-        switch (this.config("deploy.mode")) {
-            case 'gears-cli': 
-                return `gears runtime undeploy ${path}`
-            case 'maven':
-                return 'mvn org.wildfly.plugins:wildfly-maven-plugin:2.0.2.Final:undeploy'
-            default:
-                const docker = this.config('docker')
-                return `${docker} exec -it gears-runtime rm /camunda/standalone/deployments/${file}`
-        }
+    stopExecution(): Execution {
+        const projectName = this.config('project.name')
+        const cwd = `${this.workspaceRoot}/target/${projectName}`
+        const cmd = 'mvn spring-boot:stop -Dspring-boot.stop.fork=true'
+        return new vscode.ShellExecution(cmd, { cwd })
     }
 
     loadDataExecution(): Execution {